+++ lizard_wbconfiguration/tasks.py
@@ -18,6 +18,8 @@ from lizard_wbconfiguration.export_dbf import DBFExporter
 from lizard_wbconfiguration.models import DBFConfiguration
 
 from lizard_task.handler import get_handler
+from lizard_task.task import task_logging
+
 
 @task()
 def import_dbf(fews_meta_info=None,
@@ -49,6 +51,7 @@ def import_dbf(fews_meta_info=None,
     logger.removeHandler(handler)
     return "<<import dbf>>"
 
+
 def run_importdbf_task():
     """Run task import_dbf.
 
@@ -61,7 +64,21 @@ def run_importdbf_task():
               "structures_filepath": "/tmp/pumpingstations.dbf"}
     import_dbf.delay(**kwargs)
 
+
+def remove_rejected_configurations(data_set, configtype, logger):
+        """Remove rejected configurations from ConfigurationToValidate."""
+        options = {"data_set__name": data_set,
+                   "config_type": configtype,
+                   "action": ConfigurationToValidate.REJECT}
+        count = ConfigurationToValidate.objects.filter(**options).count()
+        logger.info("%d rejected configuration(s) to delete." % count)
+        if count > 0:
+            ConfigurationToValidate.objects.filter(**options).delete()
+            logger.info("%d configuration(s) deleted." % count)
+
+
 @task()
+@task_logging
 def validate_wbconfigurations(taskname="",
                               username=None,
                               levelno=20,
@@ -71,19 +88,19 @@ def validate_wbconfigurations(taskname="",
     Import wb areaconfigurations from dbf using
     validation configurations.
     """
-    action = 1
+    logger = logging.getLogger(taskname)
+    logger.info("Start validation of wbconfigurations for '%s'." % data_set)
+
+    remove_rejected_configurations(data_set, configtype, logger)
+
     v_configs = ConfigurationToValidate.objects.filter(
         data_set__name__iexact=data_set,
         config_type=configtype,
-        action=action)
+        action=ConfigurationToValidate.VALIDATE)
     v_configs = v_configs.exclude(file_path=None)
-    handler = get_handler(taskname=taskname, username=username)
-    logger = logging.getLogger(taskname)
-    logger.addHandler(handler)
-    logger.setLevel(int(levelno))
+
     validated = 0
     failed = 0
-    logger.info("Start validation of wbconfigurations for '%s'." % data_set)
     for v_config in v_configs:
         dbfimporter = DBFImporter(logger)
         dbfimporter.fews_meta_info = v_config.fews_meta_info
@@ -104,7 +121,7 @@ def validate_wbconfigurations(taskname="",
             validated = validated + 1
             logger.debug("Validated with SUCCESS.")
         else:
-            v_config.action = 0
+            v_config.action = ConfigurationToValidate.KEEP
             if isinstance(status, tuple) and len(status) > 1:
                 v_config.action_log = status[1][:256]
             else:
@@ -114,8 +131,6 @@ def validate_wbconfigurations(taskname="",
             logger.debug("Validated with ERRORS.")
     logger.info("Succeed=%s, Failed=%s." % (validated, failed))
     logger.info("End validation.")
-    logger.removeHandler(handler)
-    return "<<validate_wbconfigurations>>"
 
 
 @task()
@@ -153,11 +168,12 @@ def validate_all(taskname='validate_all', username=None):
 
 
 @task()
+@task_logging
 def export_wbconfigurations_to_dbf(
     data_set=None,
     levelno=20,
     username=None,
-    taskname="wb_export_to_dbf_all"):
+    taskname=""):
     """
     Export water balance configurations into dbf.
     Use logging handler of lizard_task app. to write message into database.
@@ -166,10 +182,7 @@ def export_wbconfigurations_to_dbf(
     data_set -- name of organisation as DataSet in lizard_security
     levelno -- logging level as number, 10=debug, 20=info, ...
     """
-    handler = get_handler(taskname=taskname, username=username)
     logger = logging.getLogger(taskname)
-    logger.addHandler(handler)
-    logger.setLevel(int(levelno))
     dbfexporter = DBFExporter(logger)
     dbf_configurations = DBFConfiguration.objects.exclude(dbf_type='Area')
     if data_set is not None:
@@ -190,44 +203,33 @@ def export_wbconfigurations_to_dbf(
         else:
             logger.debug("UNKNOWN source %s" % dbf_configuration.dbf_type)
     logger.info("END EXPORT.")
-    logger.removeHandler(handler)
 
 
 @task()
+@task_logging
 def export_aanafvoergebieden_to_dbf(
     data_set=None,
-    taskname='aanafvoergebieden_export_to_dbf_all',
+    taskname='',
     levelno=20,
     username=None):
     """
     Export geo info of 'aanafvoergebieden' into dbf.
     """
 
-    handler = get_handler(taskname=taskname, username=username)
     logger = logging.getLogger(taskname)
-    logger.addHandler(handler)
-    logger.setLevel(int(levelno))
-
-    logger.info(data_set)
-    logger.info(taskname)
-    logger.info(username)
-    logger.info(levelno)
-
     dbfexporter = DBFExporter(logger)
+    logger.info("Sart export of 'aanafvoergebieden'.")
     dbf_configurations = DBFConfiguration.objects.filter(dbf_type='Area')
     if data_set is not None:
         dbf_configurations = dbf_configurations.filter(data_set__name=data_set)
-    logger.info("%s water balance configurations to export." % len(
-            dbf_configurations))
     for dbf_configuration in dbf_configurations:
+        logger.info("Start export of 'aanafvoergebieden' for '%s'." % dbf_configuration.data_set.name)
         owner = dbf_configuration.data_set
         save_to = dbf_configuration.save_to
         filename = dbf_configuration.filename
         if dbf_configuration.dbf_type == 'Area':
             dbfexporter.export_aanafvoergebieden(owner, save_to, filename)
-
-    logger.info("Export water balance configurations is finished.")
-    logger.removeHandler(handler)
+    logger.info("END EXPORT.")
 
 
 @task()